add edinet fetching

+++ scripts/fetch_data_from_edinet.py
@@ -6,6 +6,8 @@ import zipfile
 from io import BytesIO
 from pathlib import Path
 
+from dateutil.relativedelta import relativedelta
+
 import data_fetcher
 
 api_key = "c528ad6f91db40468bf86c3f080daaff"
@@ -15,7 +17,7 @@ session = data_fetcher.session.get_session(max_requests_per_second=1)
 
 target_summary_taxonomy = dict(
     net_sales="jpcrp_cor:NetSalesSummaryOfBusinessResults",
-    ordinary_incom="jpcrp_cor:OrdinaryIncomeLossSummaryOfBusinessResults",
+    ordinary_income="jpcrp_cor:OrdinaryIncomeLossSummaryOfBusinessResults",
     net_income="jpcrp_cor:NetIncomeLossSummaryOfBusinessResults",
     comprehensive_income="jpcrp_cor:NetIncomeLossSummaryOfBusinessResults",
     net_asset="jpcrp_cor:NetAssetsSummaryOfBusinessResults",
@@ -31,7 +33,6 @@ target_summary_taxonomy = dict(
     cost_of_sales="jppfs_cor:CostOfSales",
     gross_profit="jppfs_cor:GrossProfit",
     operating_income="jppfs_cor:OperatingIncome",
-    ordinary_income="jppfs_cor:OrdinaryIncome",
 )
 
 
@@ -57,101 +58,44 @@ def get_document(doc_id):
     return rows
 
 
-def extract_data(rows, target_keys):
-    data = {}
+def append_date_period(
+    row: [str], start_date: datetime.datetime, end_date: datetime.datetime
+):
+    period_str = row[0]
+    if "Q" not in period_str and "Y" not in period_str:
+        raise ValueError(f"Invalid period string : {period_str}")
+
+    split_char = "Q" if "Q" in period_str else "Y"
+    delta = 3 if split_char == "Q" else 12
+
+    key = period_str.split(split_char)[0]
+    if key == "Current":
+        sdate, edate = start_date, end_date
+    else:
+        delta_month = relativedelta(months=int(key.replace("Prior", "")) * delta)
+        sdate = start_date - delta_month
+        edate = end_date - delta_month
+    return [sdate.strftime("%Y-%m-%d"), edate.strftime("%Y-%m-%d")] + row[1:]
+
+
+def extract_data(
+    rows: list[list[str]],
+    target_keys: dict[str, str],
+    start_date: datetime.datetime,
+    end_date: datetime.datetime,
+) -> list[list[str]]:
+    data = []
     for key, value in target_keys.items():
         extracted = [row for row in rows if row[0] == value]
-        consolidated = [d for d in extracted if "_NonConsolidatedMember" not in d[2]]
+        consolidated = [d for d in extracted if "_" not in d[2]]
         if len(consolidated) > 0:
             extracted = consolidated
 
-        extracted = [
-            [d[2].replace("_NonConsolidatedMember", ""), d[-1]] for d in extracted
+        extracted = [[d[2].split("_")[0]] + [d[0], d[2], d[-1]] for d in extracted]
+        data += [
+            [key] + append_date_period(row, start_date, end_date) for row in extracted
         ]
-        data[key] = extracted
     return data
-
-
-def create_csv_rows(
-    start_date: datetime.datetime,
-    end_date: datetime.datetime,
-    data: dict[str, list[list[str]]],
-):
-    q_data = {}
-    y_data = {}
-
-    for key, val in data.items():
-        if len(val) > 0:
-            if "Q" in val[0]:
-                q_data[key] = val
-            elif "Y" in val[0]:
-                y_data[key] = val
-
-    rows = []
-    if len(q_data) > 0:
-        rows += create_csv_rows_impl(start_date, end_date, q_data, 3)
-    if len(y_data) > 0:
-        rows += create_csv_rows_impl(start_date, end_date, y_data, 12)
-    return rows
-
-
-def create_csv_rows_impl(
-    start_date: datetime.datetime,
-    end_date: datetime.datetime,
-    data: dict[str, list[list[str]]],
-    delta: int,
-):
-    header = [key for key in data.keys()]
-
-    def split(d: list[str]):
-        split_str = "Q" if delta == 3 else "Y"
-        splits = d[0].split(split_str)
-        return splits
-
-    keys = set(sum([[split(d)[0] for d in val] for val in data.values()], []))
-    dates_dict = {}
-    for key in keys:
-        if key == "Current":
-            dates_dict[key] = [start_date, end_date]
-        else:
-            delta_month = int(key.replace("Prior", "")) * delta
-            dyear = delta_month // 12
-            dmonth = delta_month % 12
-
-            if delta_month >= start_date.month:
-                sdyear = dyear + 1
-                sdmonth = dmonth - 12
-            else:
-                sdyear = dyear
-                sdmonth = dmonth
-            sdate = datetime.datetime(
-                start_date.year - sdyear, start_date.month - sdmonth, start_date.day
-            )
-
-            if delta_month >= end_date.month:
-                edyear = dyear + 1
-                edmonth = dmonth - 12
-            else:
-                edyear = dyear
-                edmonth = dmonth
-            edate = datetime.datetime(
-                end_date.year - edyear, end_date.month - edmonth, end_date.day
-            )
-            dates_dict[key] = [sdate, edate]
-
-    rows = []
-    for key in keys:
-        row = [d.strftime("%Y-%m-%d") for d in dates_dict[key]]
-        for head in header:
-            for d in data[head]:
-                if key in d[0]:
-                    if d[1] == "－":
-                        row.append("")
-                    else:
-                        row.append(d[1])
-                    break
-        rows.append(row)
-    return rows
 
 
 def get_target_docs_info(target_date: datetime.date):
@@ -178,23 +122,65 @@ def get_target_docs_info(target_date: datetime.date):
     return target_docs
 
 
-def get_parent_root_doc_info(doc_id: str | None, target_docs: list[dict] | None):
+def get_parent_root_doc_info(doc_id: str | None, doc_list: list[list[str]]):
     if doc_id is None:
         return None
 
-    if target_docs is None:
-        return None
-
-    for doc in target_docs:
-        if doc["docID"] == doc_id:
-            if doc["parentDocID"] is not None:
-                return get_parent_root_doc_info(doc["parentDocID"], target_docs)
+    for doc in doc_list:
+        if doc[0] == doc_id:
+            if doc[-1] is not None:
+                return get_parent_root_doc_info(doc[-1], doc_list)
             return doc
     return None
 
 
+def update_document_list(target_docs: list[dict], output_dir: Path):
+    # 取得したドキュメントの情報をCSVに保存
+    doc_list_path = output_dir / "doc_list.csv"
+
+    doc_keys = [
+        "docID",
+        "edinetCode",
+        "secCode",
+        "submitDateTime",
+        "periodStart",
+        "periodEnd",
+        "parentDocID",
+    ]
+    doc_info = [[doc[key] for key in doc_keys] for doc in target_docs]
+
+    rows = []
+    if doc_list_path.exists():
+        with open(doc_list_path, "r") as f:
+            csv_reader = csv.reader(f)
+            rows = list(csv_reader)
+    else:
+        with open(doc_list_path, "w") as f:
+            csv_writer = csv.writer(f, lineterminator="\n")
+            csv_writer.writerow(doc_keys)
+
+    # 既存のドキュメント情報と新しいドキュメント情報をマージ
+    ids = [doc[0] for doc in rows]
+    new_rows = [info for info in doc_info if info[0] not in ids]
+    with open(doc_list_path, "a") as f:
+        csv_writer = csv.writer(f, lineterminator="\n")
+        csv_writer.writerows(new_rows)
+
+    return rows + new_rows
+
+
 def main(target_date: datetime.date, output_dir: Path):
+    header = [
+        "key",
+        "start_date",
+        "end_date",
+        "edinet_key",
+        "period",
+        "value",
+        "announce_date",
+    ]
     target_docs = get_target_docs_info(target_date)
+    doc_list = update_document_list(target_docs, output_dir)
 
     for doc in target_docs:
         doc_id = doc["docID"]
@@ -206,23 +192,47 @@ def main(target_date: datetime.date, output_dir: Path):
         code = doc["secCode"] or doc["edinetCode"]
 
         if doc["periodStart"] is None or doc["periodEnd"] is None:
-            parent_doc = get_parent_root_doc_info(doc["parentDocID"], target_docs)
-            doc["periodStart"] = doc["periodStart"] or parent_doc["periodStart"]
-            doc["periodEnd"] = doc["periodEnd"] or parent_doc["periodEnd"]
+            parent_doc = get_parent_root_doc_info(doc["parentDocID"], doc_list)
+            if parent_doc is not None:
+                doc["periodStart"] = doc["periodStart"] or parent_doc[4]
+                doc["periodEnd"] = doc["periodEnd"] or parent_doc[5]
+        if doc["periodStart"] is None or doc["periodEnd"] is None:
+            print("No period start or end date")
+            continue
+
         start_date = datetime.datetime.strptime(doc["periodStart"], "%Y-%m-%d")
         end_date = datetime.datetime.strptime(doc["periodEnd"], "%Y-%m-%d")
 
-        output_path = output_dir / f"{code}_{date_str}.csv"
-        data = extract_data(rows, target_summary_taxonomy)
-        header = [key for key in data.keys()]
-        rows = create_csv_rows(start_date, end_date, data)
-        rows = sorted(rows, key=lambda x: x[0])
+        output_path = output_dir / f"{code}.csv"
+        data = [
+            row + [date_str]
+            for row in extract_data(rows, target_summary_taxonomy, start_date, end_date)
+        ]
 
         if len(rows) > 0:
-            with open(output_path, "w", encoding="utf-8") as f:
-                csv_writer = csv.writer(f, lineterminator="\n")
-                csv_writer.writerow(["start_date", "end_date"] + header)
-                csv_writer.writerows(rows)
+            if output_path.exists():
+                with open(output_path, "r", encoding="utf-8") as f:
+                    csv_reader = csv.reader(f)
+                    header_row = next(csv_reader)
+                    rows = list(csv_reader)
+
+                # exclude duplicate
+                filtered = []
+                for d in data:
+                    for row in rows:
+                        if d[3] == row[3] and d[4] == row[4] and d[5] == row[5]:
+                            break
+                    else:
+                        filtered.append(d)
+
+                with open(output_path, "a", encoding="utf-8") as f:
+                    csv_writer = csv.writer(f, lineterminator="\n")
+                    csv_writer.writerows(filtered)
+            else:
+                with open(output_path, "w", encoding="utf-8") as f:
+                    csv_writer = csv.writer(f, lineterminator="\n")
+                    csv_writer.writerow(header)
+                    csv_writer.writerows(data)
             print("Saved : ", output_path)
